refactor dup vars
