+++ src/index/src/inverted_index.rs
@@ -18,5 +18,3 @@ pub mod format;
 pub mod search;
 
 pub type FstMap = fst::Map<Vec<u8>>;
-pub type Bytes = Vec<u8>;
-pub type BytesRef<'a> = &'a [u8];
