+++ src/index/src/lib.rs
@@ -20,3 +20,6 @@ pub mod error;
 pub mod external_provider;
 pub mod fulltext_index;
 pub mod inverted_index;
+
+pub type Bytes = Vec<u8>;
+pub type BytesRef<'a> = &'a [u8];
