+++ src/mito2/src/sst/index/inverted_index/applier/builder/comparison.rs
@@ -14,7 +14,7 @@
 
 use datafusion_expr::{Expr as DfExpr, Operator};
 use index::inverted_index::search::predicate::{Bound, Predicate, Range, RangePredicate};
-use index::inverted_index::Bytes;
+use index::Bytes;
 
 use crate::error::Result;
 use crate::sst::index::inverted_index::applier::builder::InvertedIndexApplierBuilder;
